+++ src/fragmentIterators/MergeFragments.h
@@ -1,8 +1,4 @@
 #pragma once
-
-#include "../lib/dary_heap.hpp"
-#include <cstring>
-#include <unordered_map>
 #include <vector>
 
 #include "FragmentIterator.h"
@@ -16,30 +12,64 @@ namespace BPCells {
 // All inputs must have known cell counts, chr counts+names, and be seekable
 class MergeFragments : public FragmentLoader {
   private:
-    std::vector<FragmentIterator> frags;
-    std::vector<uint32_t> heap; // Heap of indices into the fragments array
+    // Helper class to handle reading a fixed number of fragments at a time,
+    // but in bulk.
+    class ChunkedLoader {
+        std::unique_ptr<FragmentLoader> frags;
+        uint32_t loaded = 0; // How many fragments are currently loaded and available to copy
+        const uint32_t cell_offset, chunk_size; 
+
+      public:
+        ChunkedLoader(std::unique_ptr<FragmentLoader> &&loader, uint32_t cell_offset, uint32_t chunk_size);
+        // Copy fragments to the destinations (must have space for at least chunk_size elements)
+        // Return how many fragments were actually copied
+        uint32_t load_chunk(uint32_t *start, uint32_t *end, uint32_t *cell);
+        
+        // Return the next start coordinate without marking it loaded, or UINT32_MAX if no more available
+        uint32_t peek_start(); 
+
+        // Wrapper methods for internal fragments
+        void seek(uint32_t chr_id, uint32_t base);
+        void restart();
+
+        int chrCount() const;
+        int cellCount() const;
+
+        const char *chrNames(uint32_t chr_id);
+        const char *cellNames(uint32_t cell_id);
+
+        bool nextChr();
+
+        uint32_t currentChr() const;
+    };
+
+    uint32_t const load_size, chunk_size, chunks_per_input;
+
+    std::vector<ChunkedLoader> frags;
     std::vector<uint32_t> cell_id_offset;
 
     std::vector<std::string> chr_order; // Order of output chromosomes
     std::vector<std::vector<uint32_t>>
         source_chr; // source_chr[i][j] is the chr ID in frags[i] with name chr_order[j]
 
-    std::vector<uint32_t> start, end, cell;
-    uint32_t loaded;
-    uint32_t current_chr = UINT32_MAX;
+    // Size these to 3 * chunk_size * frags.size() 
+    std::vector<uint32_t> start, end, cell, start_buf, end_buf, cell_buf;
 
-    // Order by start, breaking ties by the fragment loader index
-    inline bool compare(uint32_t idx1, uint32_t idx2) {
-        if (frags[idx1].start() != frags[idx2].start())
-            return frags[idx1].start() > frags[idx2].start();
-        return idx1 > idx2;
-    }
+    // Heap of (start_coord, loader_idx) pairs to use in load() operation
+    std::vector<std::pair<uint32_t, uint32_t>> heap;
+
+    // Loaded is how many values are saved in the data buffers
+    // Available is how many of those are ready to output
+    // cur_idx is the current index that we are outputting
+    uint32_t loaded, available, cur_idx; 
+    uint32_t current_chr = UINT32_MAX;
 
   public:
     MergeFragments(
         std::vector<std::unique_ptr<FragmentLoader>> &&fragments,
         const std::vector<std::string> &chr_order,
-        uint32_t load_size = 1024
+        uint32_t load_size = 1024, // Output load size
+        uint32_t chunk_size = 32 // Input load chunk size
     );
 
     ~MergeFragments() = default;
@@ -58,6 +88,16 @@ class MergeFragments : public FragmentLoader {
 
     uint32_t currentChr() const override;
 
+    // Load algorithm:
+    // 1. Add chunks_per_input * N chunks from the N input loaders into the cell, start, end vectors
+    //    - Always add the chunk with the smallest start coordinate
+    //    - Track what's the largest start coordinate added from any input loader.
+    //      The smallest of these coordinates well be the limit of our known-sorted data.
+    // 2. Sort all the fragments
+    // 3. Output fragments until we hit our start coordinate limit
+    //    - This is guaranteed to leave at most N chunks left over (at most 1 chunk for each input loader)
+    //    - This means if we have space for (chunks_per_input + 1)*N chunks, and add (chunks_per_input)*N chunks each time we can never overflow 
+    //      our buffers, and we'll never sort a fragment more than once
     bool load() override;
 
     uint32_t capacity() const override;